Google Image Scraper
By changing the variable "images_of", this script will create a new folder with images of the search query. This was built with the purpose to ease data collecting for model building. After which you can scrape through to delete unwanted images and easily label them. Still a work in progress.